% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver}
\alias{liver}
\title{Liver trajectory example data}
\format{A `list` object.}
\usage{
data(liver, package = 'scHOT')
}
\description{
A liver data set contains two branches (hepatocyte and cholangiocyte).
}
\keyword{datasets}
