% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOB_subset}
\alias{MOB_subset}
\title{MOB_subset spatial example data}
\format{
An object of class \code{list} of length 1.
}
\usage{
data(MOB_subset, package = 'scHOT')
}
\description{
A MOB spatial data set.
}
\keyword{datasets}
