% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_plot.R
\name{plotEgoNetwork}
\alias{plotEgoNetwork}
\title{plotEgoNetwork}
\usage{
plotEgoNetwork(
  scHOT,
  hubnode,
  network,
  weight = "higherOrderStatistic",
  subset = FALSE,
  thresh = NULL
)
}
\arguments{
\item{scHOT}{a scHOT object}

\item{hubnode}{is a character vector of node(s) to include as hub nodes}

\item{network}{is an igraph network}

\item{weight}{A string indicates the column name stored in scHOT_output slot that are used as the weights of the network}

\item{subset}{is a logical asking if you should subset based on the weight (default FALSE)}

\item{thresh}{is the subset weight threshold}
}
\value{
\code{igraph} object containing the network graphed. Produces an igraph plot
}
\description{
the plotEgoNetwork function plots network graphs with edges coloured by weights in the network
}
