% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_class.R
\docType{class}
\name{scHOT-class}
\alias{scHOT-class}
\title{scHOT class}
\description{
scHOT class
}
\section{Slots}{

\describe{
\item{\code{testingScaffold}}{A matrix of scaffold}

\item{\code{weightMatrix}}{A matrix or dgCMatrix indicates the weight matrix}

\item{\code{scHOT_output}}{A data.frame or DtatFrame to store output from scHOT}

\item{\code{params}}{A list of parameters}
}}

