% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_buildFromMatrix}
\alias{scHOT_buildFromMatrix}
\title{Create scHOT object from a matrix}
\usage{
scHOT_buildFromMatrix(mat, cellData = NULL)
}
\arguments{
\item{mat}{A matrix with rows for genes and columns for cells}

\item{cellData}{A dataframe or DataFrame object with rows for cells}
}
\description{
Create scHOT object from a matrix
}
