% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_output<-}
\alias{scHOT_output<-}
\title{Setter functions for scHOT objects}
\usage{
scHOT_output(x) <- value
}
\arguments{
\item{x}{A scHOT object}

\item{value}{The value the slot should take}
}
\value{
A scHOT object
}
\description{
Setter functions for scHOT objects
}
\examples{

data(MOB_subset)
sce_MOB_subset <- MOB_subset$sce_MOB_subset
scHOT_spatial <- scHOT_buildFromSCE(sce_MOB_subset,
assayName = "logcounts",
positionType = "spatial",
positionColData = c("x", "y"))

scHOT_output(scHOT_spatial) <- DataFrame()

}
