% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_plot.R
\name{scHOT_plotPermutationDistributions}
\alias{scHOT_plotPermutationDistributions}
\title{scHOT_plotPermutationDistributions}
\usage{
scHOT_plotPermutationDistributions(scHOT)
}
\arguments{
\item{scHOT}{a scHOT object}
}
\value{
\code{ggplot} graph of global higher order function and the
permutation scHOT test statistics. This should have a continuous pattern
to be reliably used for p-value estimation
}
\description{
the scHOT_plotPermutationDistributions function plots
the permutation test statistics as a diagnostic plot
for estimating p-values
}
