% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_setPermutationScaffold}
\alias{scHOT_setPermutationScaffold}
\title{scHOT_setPermutationScaffold}
\usage{
scHOT_setPermutationScaffold(
  scHOT,
  numberPermutations = 1000,
  numberScaffold = 100,
  storePermutations = TRUE
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{numberPermutations}{The number of permutations, set as 1000 by default}

\item{numberScaffold}{The number of scaffold, set as 100 by default. if you want all combinations to do permutations then set, numberScaffold much higher than the testingScaffold}

\item{storePermutations}{a logical flag on whether
Permutations should be stored, or discarded once used}
}
\description{
Set permutation scaffold
}
