% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_setWeightMatrix}
\alias{scHOT_setWeightMatrix}
\title{scHOT_setWeightMatrix}
\usage{
scHOT_setWeightMatrix(
  scHOT,
  weightMatrix = NULL,
  positionType = "trajectory",
  positionColData = NULL,
  nrow.out = NULL,
  averageAcrossTrajectoryTies = FALSE,
  ...
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{weightMatrix}{A matrix indicating the weight matrix for scHOT analysis, such as the output from `trajectoryWeightMatrix` or `spatialWeightMatrix`. If this is not NULL then other parameters are ignored.}

\item{positionType}{A string indicating the position type, either "trajectory" or "spatial"}

\item{positionColData}{Either trajectory or spatial information for each sample. If positionType is "trajectory" then positionColData should be a character or numeric indicating the subset of colData of the scHOT object. If positionType is "spatial" then positionColData should be a character or numeric vector indicating the subset of colData that give the full spatial coordinates.}

\item{nrow.out}{The number of weightings to include for testing, a smaller value is faster for computation}

\item{averageAcrossTrajectoryTies}{Logical indicating whether ties in the trajectory should be given the same local weights}

\item{...}{parameters for function trajectoryWeightMatrix or spatialWeightMatrix}
}
\description{
Create scHOT object from a SingleCellExperiment object
}
