% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{testingScaffold<-}
\alias{testingScaffold<-}
\title{Setter functions for scHOT objects}
\usage{
testingScaffold(x) <- value
}
\arguments{
\item{x}{A scHOT object}

\item{value}{The value the slot should take}
}
\value{
A scHOT object
}
\description{
Setter functions for scHOT objects
}
\examples{

data(MOB_subset)
sce_MOB_subset <- MOB_subset$sce_MOB_subset
scHOT_spatial <- scHOT_buildFromSCE(sce_MOB_subset,
assayName = "logcounts",
positionType = "spatial",
positionColData = c("x", "y"))

pairs <- t(combn(rownames(sce_MOB_subset),2))
rownames(pairs) <- apply(pairs,1,paste0,collapse = "_")

testingScaffold(scHOT_spatial) <- pairs

}
