% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{trajectoryWeightMatrix}
\alias{trajectoryWeightMatrix}
\title{trajectoryWeightMatrix}
\usage{
trajectoryWeightMatrix(n, type = NULL, span = NULL)
}
\arguments{
\item{n}{indicates the number of cels}

\item{type}{Type of weight matrix, one of "harmonic", "triangular", "block"}

\item{span}{proportion of samples to include on either side, default is 0.5}
}
\description{
Create weight matrix for trajectory data
}
\examples{

trajectoryWeightMatrix(100)
trajectoryWeightMatrix(100, type = "triangular")
trajectoryWeightMatrix(100, type = "block")
trajectoryWeightMatrix(100, type = "harmonic")

}
