% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{weightMatrix<-}
\alias{weightMatrix<-}
\title{Setter functions for scHOT objects}
\usage{
weightMatrix(x) <- value
}
\arguments{
\item{x}{A scHOT object}

\item{value}{The value the slot should take}
}
\value{
A scHOT object
}
\description{
Setter functions for scHOT objects
}
\examples{

library(SingleCellExperiment)

data(MOB_subset)
sce_MOB_subset <- MOB_subset$sce_MOB_subset
scHOT_spatial <- scHOT_buildFromSCE(sce_MOB_subset,
assayName = "logcounts",
positionType = "spatial",
positionColData = c("x", "y"))

W <- spatialWeightMatrix(colData(scHOT_spatial)[,slot(scHOT_spatial, "positionColData")])

weightMatrix(scHOT_spatial) <- W

}
