% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_plot.R
\name{plotColouredExpression}
\alias{plotColouredExpression}
\title{plotColouredExpression}
\usage{
plotColouredExpression(
  scHOT,
  genepair,
  branches = NULL,
  ranked_by = NULL,
  subsetBranch = NULL,
  n = 3,
  fittedline = TRUE
)
}
\arguments{
\item{scHOT}{A scHOT object, where the expression data is stored in the assay slot, with assay name "expression".}

\item{genepair}{is either a single character string with an underscore, or a two length character vector}

\item{branches}{A character indicates that the colnames stored the branch information in colData}

\item{ranked_by}{A character indicates that the colnames stored the ranking information of the cells in colData, such as trajectory time,
If it is NULL, it will be ranked based on the branch information.}

\item{subsetBranch}{subsetBranch is a character vector containing the names of the branches to be plotted. If NULL it will plot all branches}

\item{n}{number of panels to split ranked samples into, default 3.}

\item{fittedline}{logical default TRUE, add a lm straight line to the plot}
}
\value{
\code{ggplot} a ggplot object of scatterplots of expression split by sample ordering
}
\description{
the plotColouredExpression function plots an n-panel scatterplot of the gene pairs split by early, mid, and late in the sample ordering.
}
