% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_plot.R
\name{plotHigherOrderSequence}
\alias{plotHigherOrderSequence}
\title{plotHigherOrderSequence}
\usage{
plotHigherOrderSequence(scHOT, gene, branches = NULL)
}
\arguments{
\item{scHOT}{A scHOT object with higherOrderSequence in scHOT_output slot}

\item{gene}{is either a logical vector matching rows of entries in wcorsList, or a character of a gene}

\item{branches}{A character indicates that the colnames stored the branch information in colData}
}
\value{
\code{ggplot} object with line plots
}
\description{
the plotHigherOrderSequence function plots weighted correlation vectors (stored in higherOrderSequence) as line plots
}
