% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_class.R
\docType{class}
\name{scHOT-class}
\alias{scHOT-class}
\title{scHOT class}
\description{
scHOT class
}
\section{Slots}{

\describe{
\item{\code{testingScaffold}}{A matrix with rows for each testing combination}

\item{\code{weightMatrix}}{A matrix or dgCMatrix indicates the weight matrix}

\item{\code{scHOT_output}}{A data.frame or DtatFrame to store output from scHOT}

\item{\code{params}}{A list of parameters}

\item{\code{positionType}}{A character indicates the type of the position,
either trajectory or spatial}

\item{\code{positionColData}}{A vector indicates column names of colData
that stored the postion informaton}
}}

