% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT.R
\name{scHOT}
\alias{scHOT}
\title{scHOT}
\usage{
scHOT(
  scHOT,
  testingScaffold = NULL,
  weightMatrix = NULL,
  positionType = NULL,
  positionColData = NULL,
  nrow.out = NULL,
  averageAcrossTrajectoryTies = FALSE,
  higherOrderFunction = NULL,
  higherOrderFunctionType = NULL,
  numberPermutations = 1000,
  numberScaffold = 100,
  storePermutations = TRUE,
  higherOrderSummaryFunction = sd,
  parallel = FALSE,
  BPPARAM = BiocParallel::SerialParam(),
  usenperm = FALSE,
  nperm = 10000,
  maxDist = 0.1,
  plot = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{testingScaffold}{A matrix with rows for each testing combination}

\item{weightMatrix}{A matrix indicates the weight matrix for scHOT analysis}

\item{positionType}{A string indicates the position type, either trajectory or spatial}

\item{positionColData}{If positionType is "trajectory" then positionColData should be a sortable vector
if positionType is "spatial" then positionColData should be a matrix type object.
It should be stored in colData (SHILA to check).}

\item{nrow.out}{SHILA to input}

\item{averageAcrossTrajectoryTies}{SHILA to input}

\item{higherOrderFunction}{A function object indicates the higher order function}

\item{higherOrderFunctionType}{is weighted or unweighted, determines if there
is a weighting argument in the higher order function}

\item{numberPermutations}{The number of permutatuion, set as 1000 by default}

\item{numberScaffold}{The number of scaffold, set as 100 by default}

\item{storePermutations}{a logical flag on whether}

\item{higherOrderSummaryFunction}{A functon indicates the higher order summary function (SHILA to check)}

\item{parallel}{A logical input indicates whether run the permutation test using multiple cores in parallel.}

\item{BPPARAM}{A \code{BiocParallelParam} class object from the \code{BiocParallel} package is used. Default is SerialParam().}

\item{usenperm}{SHILA to write}

\item{nperm}{number of permutation}

\item{maxDist}{SHILA to write}

\item{plot}{A logical input indicates whether the results are plotted}

\item{verbose}{A logical input indicates whether the intermediate steps will be printed}

\item{...}{parameters for function trajectoryWeightMatrix or spatialWeightMatrix}
}
\description{
A wrap up function to perform scHOT
}
