% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_addTestingScaffold}
\alias{scHOT_addTestingScaffold}
\title{scHOT_addTestingScaffold}
\usage{
scHOT_addTestingScaffold(scHOT, testingScaffold)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{testingScaffold}{A matrix with rows for each testing combination, and columns for level of dimensionality (1 for single gene etc.)}
}
\description{
Add a testing scaffold to a scHOT object
}
