% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_buildFromSCE}
\alias{scHOT_buildFromSCE}
\title{Create scHOT object from a SingleCellExperiment object}
\usage{
scHOT_buildFromSCE(sce, assayName = "counts")
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{assayName}{is a single assay to pull out from sce}
}
\description{
Create scHOT object from a SingleCellExperiment object
}
