% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_buildFromSCE}
\alias{scHOT_buildFromSCE}
\title{scHOT_buildFromSCE}
\usage{
scHOT_buildFromSCE(sce, assayName = "counts")
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{assayName}{is a single assay to pull out from sce}
}
\description{
Create scHOT object from a SingleCellExperiment object
}
\examples{

library(SingleCellExperiment)
dat <- rbind(rnorm(50), rnorm(50), rnorm(50))
colnames(dat) <- paste0("cell_", 1:ncol(dat))
rownames(dat) <- c("T","Gata1", "Tal1")

sce <- SingleCellExperiment::SingleCellExperiment(assays = S4Vectors::SimpleList(counts = dat))
scHOT <- scHOT_buildFromSCE(sce)

}
