% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_calculateGlobalHigherOrderFunction}
\alias{scHOT_calculateGlobalHigherOrderFunction}
\title{scHOT_calculateGlobalHigherOrderFunction}
\usage{
scHOT_calculateGlobalHigherOrderFunction(
  scHOT,
  higherOrderFunction = NULL,
  higherOrderFunctionType = NULL
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{higherOrderFunction}{A function object indicates the higher order function}

\item{higherOrderFunctionType}{is weighted or unweighted, determines if there
is a weighting argument in the higher order function}
}
\description{
this calculates the global higher order function and stores it in the output
if these aren't found in the params slot then they need to be specified here
}
\details{
Calculates the global higher order function
}
