% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_functions.R
\name{scHOT_calculateHigherOrderTestStatistics}
\alias{scHOT_calculateHigherOrderTestStatistics}
\title{scHOT_calculateHigherOrderTestStatistics}
\usage{
scHOT_calculateHigherOrderTestStatistics(
  scHOT,
  higherOrderSummaryFunction = NULL,
  ...
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{higherOrderSummaryFunction}{A functon indicates the higher order summary function (SHILA to check)}

\item{...}{parameters for higherOrderSummaryFunction}
}
\description{
Calculate and store the higherOrderSequence and higherOrderTestStatistic
}
