% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_functions.R
\name{scHOT_calculateHigherOrderTestStatistics}
\alias{scHOT_calculateHigherOrderTestStatistics}
\title{scHOT_calculateHigherOrderTestStatistics}
\usage{
scHOT_calculateHigherOrderTestStatistics(
  scHOT,
  higherOrderSummaryFunction = NULL,
  ...
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{higherOrderSummaryFunction}{A function which indicates how the higher
order sequence is summarised, default is sd}

\item{...}{parameters for higherOrderSummaryFunction}
}
\value{
scHOT A scHOT object with results stored in scHOT_output slot
}
\description{
Calculate and store the higherOrderSequence and higherOrderTestStatistic
}
