% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_functions.R
\name{scHOT_estimatePvalues}
\alias{scHOT_estimatePvalues}
\title{scHOT_estimatePvalues}
\usage{
scHOT_estimatePvalues(
  scHOT,
  usenperm = FALSE,
  nperm = 10000,
  maxDist = 0.1,
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{usenperm}{SHILA to write}

\item{nperm}{number of permutation}

\item{maxDist}{SHILA to write}

\item{plot}{A logical input indicates whether the results are plotted}

\item{verbose}{A logical input indicates whether the intermediate steps will be printed}
}
\value{
scHOT A scHOT object with results stored in scHOT_output slot
}
\description{
Estimate p-values based on already run permutation tests
}
