% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_functions.R
\name{scHOT_estimatePvalues}
\alias{scHOT_estimatePvalues}
\title{scHOT_estimatePvalues}
\usage{
scHOT_estimatePvalues(
  scHOT,
  usenperm_estimate = FALSE,
  nperm_estimate = 10000,
  maxDist = 0.1,
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{usenperm_estimate}{Logical (default FALSE) if number of neighbouring permutations should be used, or if difference of global higher order statistic should be used}

\item{nperm_estimate}{Number of neighbouring permutations to use for p-value estimation}

\item{maxDist}{max difference of global higher order statistic to use for p-value estimation (default 0.1)}

\item{plot}{A logical input indicating whether the results are plotted}

\item{verbose}{A logical input indicating whether the intermediate steps will be printed}
}
\value{
scHOT A scHOT object with results stored in scHOT_output slot
}
\description{
Estimate p-values based on already run permutation tests
}
