% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHOT_functions.R
\name{scHOT_performPermutationTest}
\alias{scHOT_performPermutationTest}
\title{scHOT_performPermutationTest}
\usage{
scHOT_performPermutationTest(
  scHOT,
  verbose = FALSE,
  parallel = FALSE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{verbose}{A logical input indicates whether the intermediate steps will be printed}

\item{parallel}{A logical input indicates whether run the permutation test using multiple cores in parallel.}

\item{BPPARAM}{A \code{BiocParallelParam} class object from the \code{BiocParallel} package is used. Default is SerialParam().}
}
\description{
Perform permutation test
}
