% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{scHOT_setWeightMatrix}
\alias{scHOT_setWeightMatrix}
\title{scHOT_setWeightMatrix}
\usage{
scHOT_setWeightMatrix(
  scHOT,
  weightMatrix = NULL,
  positionType = "trajectory",
  positionColData = NULL,
  nrow.out = NULL,
  averageAcrossTrajectoryTies = FALSE,
  ...
)
}
\arguments{
\item{scHOT}{A scHOT object}

\item{weightMatrix}{A matrix indicates the weight matrix for scHOT analysis}

\item{positionType}{A string indicates the position type, either trajectory or spatial}

\item{positionColData}{If positionType is "trajectory" then positionColData should be a sortable vector
if positionType is "spatial" then positionColData should be a matrix type object.
It should be stored in colData (SHILA to check).}

\item{nrow.out}{SHILA to input}

\item{averageAcrossTrajectoryTies}{SHILA to input}

\item{...}{parameters for function trajectoryWeightMatrix or spatialWeightMatrix}
}
\description{
Create scHOT object from a SingleCellExperiment object
}
